Move the tag/untag code for toml-test-{en,de}coder to internal/tag
This was duplicated in toml_test.go and the ./cmd/toml-test* commands.
Need to modify it to fix the locale datetime tests.

No functional changes, just moves the code.
